mailbox-create action works for GET Refactored a bit Quite a few template tweaks and optimizations

+++ lib/Ashafix/Model/Domainadmin.pm
@@ -22,6 +22,7 @@ use base 'Ashafix::Model::Base';
 our %queries = (
     select_domain_count => "SELECT count(*) FROM %table_domain_admins WHERE username=? AND domain='ALL'",
     check_global_admin  => "SELECT username FROM %table_domain_admins WHERE username=? AND domain='ALL' AND active='1'",
+    check_domain_owner  => "SELECT 1 FROM %table_domain_admins WHERE username=? AND (domain=? OR domain='ALL') AND active='1'",
     select_global_admin => "SELECT * FROM %table_domain_admins WHERE username=? AND domain='ALL'",
     insert_domadmin     => "INSERT INTO %table_domain_admins (username,domain,created) VALUES (?,?,NOW())",
     delete_by_user      => "DELETE FROM %table_domain_admins WHERE username=?",
